feat: creado componente Content para campaña escolar
